Merging to Main for deploy 2

+++ controllers/roasterController.js
@@ -1,8 +1,27 @@
 const express = require("express");
 const router = express.Router()
+const jwt = require("jsonwebtoken");
 
 const mongoose = require("mongoose");
 const Roaster = require("../models/roasterModel")
+
+const checkAuthStatus = request => {
+    if (!request.headers.authorization) {
+        return false
+    }
+    const token = request.headers.authorization.split(" ")[1]
+
+    const loggedInUser = jwt.verify(token, process.env.JWT_SECRET, (err, data) => {
+        if (err) {
+            return false
+        }
+        else {
+            return data
+        }
+    });
+    console.log(loggedInUser)
+    return loggedInUser
+}
 
 // Get all roasters
 router.get("/api/roasters", async function (req, res) {
@@ -35,15 +54,23 @@ router.get("/api/roasters/search/:name", async function (req, res) {
                     $regex: req.params.name, $options: "i"
                 }
             })
+        if (roaster.length < 1) {
+            throw("No roasters found")
+        }
         res.json(roaster)
     } catch (err) {
         console.error(err)
-        res.status(500).send("And error has appeared!")
+        res.status(404).send(err)
     }
 })
 
 // Add a roaster
 router.post("/api/roasters", async function (req, res) {
+    const loggedInUser = checkAuthStatus(req);
+    if(!loggedInUser){
+        return res.status(401).send("Must be logged in")
+    }
+    console.log(loggedInUser);
     try {
         let newRoaster = await Roaster.create({
             name: req.body.name,
@@ -60,6 +87,11 @@ router.post("/api/roasters", async function (req, res) {
 
 // Edit a roaster
 router.put("/api/roasters/:id", async function (req, res) {
+    const loggedInUser = checkAuthStatus(req);
+    if(!loggedInUser){
+        return res.status(401).send("Must be logged in")
+    }
+    console.log(loggedInUser);
     try {
         if (req.body.cafeId) {
             await Roaster.findOneAndUpdate({
@@ -91,6 +123,11 @@ router.put("/api/roasters/:id", async function (req, res) {
 
 // Delete a roaster
 router.delete("/api/roasters/:id", async function (req, res) {
+    const loggedInUser = checkAuthStatus(req);
+    if(!loggedInUser){
+        return res.status(401).send("Must be logged in")
+    }
+    console.log(loggedInUser);
     try {
         let result = await Roaster.deleteOne({ _id: mongoose.Types.ObjectId(req.params.id) })
         res.json(result)
